new dynamics working for simple accel case
new dynamics work! uses long. load transfer to calculate wheel loads. turning still disabled, comes next.
needed to balance the forces on the car by adding in normal force - sounds silly but previously had just constrained velocity in binormal direction to zero, forcing it to follow the surface of the track. but now this is done with a crude contact force. this will most certainly need to be refined if we ever go 3d. will need more matrix rotations